Renaming Adapter\ directory to ModelAdapter\.
