+++ src/Controller/OpinionController.php
@@ -9,7 +9,7 @@ use App\Form\OpinionType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use App\Repository\Adapter\EntityRepositoryInterface;
+use App\Repository\ModelAdapter\EntityRepositoryInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
